requirements_gpu.txt
tensorflow-gpu
tensorflow-probability
numpy
matplotlib
qiskit
qiskit-aer-gpu
pylatexenc
